# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Spec for footbal team management",
        "title": "API Football Team Management",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "afrizalsebastian",
            "email": "sebastiangurning@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
...
        "/api/v1/goals/{goalId}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete goal (soft-delete)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Goals"
                ],
                "summary": "Delete goal (soft-delete)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Goal ID",
                        "name": "goalId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/api.WebResponse-any"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.WebResponse-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.WebResponse-any"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.ErrorsDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "field": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "api.WebResponse-any": {
            "type": "object",
            "properties": {
                "data": {},
                "error_message": {
                    "type": "string"
                },
                "errors_detail": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.ErrorsDetail"
                    }
                },
                "http_code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "message_code": {
                    "type": "integer"
                },
                "message_id": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Please input token with format: Bearer __TOKEN__",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}